Move common tests to parent test file
